First commit
Used to create all the geometry features for the DARPA SUBOFF model
described here http://www.dtic.mil/dtic/tr/fulltext/u2/a210642.pdf

The old program was written in Fortran but cannot be found.

I have not asked for permission to publish this, but I assume it is ok
as the equations and original program code is freely avaliable.

SALOME 7.3.0 For Windows was used